+++ packages/@n8n/ai-workflow-builder.ee/evaluations/langsmith/evaluator.ts
@@ -5,7 +5,7 @@ import type { INodeTypeDescription } from 'n8n-workflow';
 
 import type { SimpleWorkflow } from '../../src/types/workflow.js';
 import { evaluateWorkflow } from '../chains/workflow-evaluator.js';
-import { programmaticEvaluation } from '../programmatic/programmatic.js';
+import { programmaticEvaluation } from '../programmatic/programmatic-evaluation';
 import type { EvaluationInput, CategoryScore } from '../types/evaluation.js';
 import {
 	isSimpleWorkflow,
