+++ packages/@n8n/ai-workflow-builder.ee/evaluations/programmatic/evaluators/agent-prompt.ts
@@ -1,7 +1,8 @@
 import type { SimpleWorkflow } from '@/types';
 import { validateAgentPrompt } from '@/validation/checks';
 import type { SingleEvaluatorResult } from '@/validation/types';
-import { calcSingleEvaluatorScore } from '@/validation/utils/score';
+
+import { calcSingleEvaluatorScore } from '../../utils/score';
 
 export function evaluateAgentPrompt(workflow: SimpleWorkflow): SingleEvaluatorResult {
 	const violations = validateAgentPrompt(workflow);
