+++ packages/@n8n/ai-workflow-builder.ee/evaluations/programmatic/evaluators/tools.ts
@@ -3,7 +3,8 @@ import type { INodeTypeDescription } from 'n8n-workflow';
 import type { SimpleWorkflow } from '@/types';
 import { validateTools } from '@/validation/checks';
 import type { SingleEvaluatorResult } from '@/validation/types';
-import { calcSingleEvaluatorScore } from '@/validation/utils/score';
+
+import { calcSingleEvaluatorScore } from '../../utils/score';
 
 export function evaluateTools(
 	workflow: SimpleWorkflow,
